pyproject.toml
[project]
name = "tomodachi-macrogen"
version = "0.1.0"
description = "Generate GLaMS/SwiCC macro scripts for Tomodachi Life face paint automation."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "Pillow>=10.0",
]

[project.scripts]
tomodachi-macrogen = "tomodachi_macrogen:main"

[dependency-groups]
dev = [
  "pyupgrade>=3.20.0",
  "ruff>=0.11.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["tomodachi_macrogen"]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
